backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
aiogram
click==8.1.7
ecdsa==0.19.0
fastapi
greenlet
h11==0.14.0
httpx
idna==3.10
passlib==1.7.4
psycopg2-binary
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn[standard]
websockets

